+++ templates/custom-indexer-template/src/components/WalletProvider.tsx
@@ -12,12 +12,6 @@ export const WalletProvider = ({ children }: PropsWithChildren) => {
     <AptosWalletAdapterProvider
       autoConnect={true}
       dappConfig={{ network: NETWORK, aptosApiKeys: {[NETWORK]: APTOS_API_KEY} }}
-      optInWallets={[
-        "Petra",
-        "Nightly",
-        "Pontem Wallet",
-        "Mizu Wallet",
-      ]}
       onError={(error) => {
         toast({
           variant: "destructive",